Brain/networking/requirements.txt
paho-mqtt
fastapi
uvicorn
aiortc
requests
jinja2